Refactor exceptions in wallet_handler
