+++ aries_cloudagent/protocols/connections/v1_0/manager.py
@@ -27,7 +27,6 @@ from ....wallet.crypto import create_keypair, seed_to_did
 from ....wallet.error import WalletNotFoundError
 from ....wallet.util import bytes_to_b58
 from ....protocols.routing.v1_0.manager import RoutingManager
-from ....wallet_handler.error import KeyNotFoundError
 
 from ....wallet_handler.handler import WalletHandler
 
@@ -131,7 +130,7 @@ class ConnectionManager:
                 wallet_handler: WalletHandler = await self.context.inject(WalletHandler)
                 try:
                     my_label = await wallet_handler.get_label(self.context)
-                except KeyNotFoundError:
+                except WalletNotFoundError:
                     pass
             else:
                 my_label = self.context.settings.get("default_label")
@@ -143,7 +142,7 @@ class ConnectionManager:
             wallet_handler: WalletHandler = await self.context.inject(WalletHandler)
             try:
                 image_url = await wallet_handler.get_image_url(self.context)
-            except KeyNotFoundError:
+            except WalletNotFoundError:
                 pass
 
         wallet: BaseWallet = await self.context.inject(BaseWallet)
@@ -349,7 +348,7 @@ class ConnectionManager:
                 wallet_handler: WalletHandler = await self.context.inject(WalletHandler)
                 try:
                     my_label = await wallet_handler.get_label(self.context)
-                except KeyNotFoundError:
+                except WalletNotFoundError:
                     pass
             else:
                 my_label = self.context.settings.get("default_label")
