+++ aries_cloudagent/wallet_handler/error.py
@@ -15,13 +15,6 @@ class WalletMissmatchError(WalletError):
     """Wrong wallet exception."""
 
 
-class WalletNotFoundError(WalletError):
-    """No wallet for given information exception."""
-
-
 class DuplicateMappingError(WalletError):
     """Mapping already exists exception."""
 
-
-class WalletDuplicateError(WalletError):
-    """Duplicate record exception."""
